* origin/master: (177 commits)
  EAMxx: add field alias unit test
  EAMxx: store alloc props and extra data via pointers in FieldHeader
  EAMxx: don't expose extra data impl from FieldHeader
  fix output list and improve action verbosity
  fix paths in docs
  improve nested lists
  minor fixes, typos, formatting
  Also adjust macmic for ne120. Changing it from 3 to 6
  switch to mpich 8.1.25
  Cray GPU/Frontier: Update flags
  custom flags for banddiag file
  update frontier time limits
  EAMxx: add more diags to model restart test output
  EAMxx: fix vapor flux unit test
  EAMxx: misc cleanup of diagnostics
  EAMxx: match long/short wave diag names to the string used to register them
  EAMxx: check that diag inputs satisfy requested pack size
  EAMxx: remove diag timestamp update from diagnostic classes
  add evp patch to machines specs for pm
  EAMxx: unify precip surf mass flux tests in one file
  ...

+++ components/eamxx/src/share/tests/eamxx_time_interpolation_tests.cpp
@@ -419,7 +419,7 @@ std::vector<std::string> create_test_data_files(
   // Jumble the order of files to also test the sorting algorithm
   auto list_of_files_tmp = om.get_list_of_files();
   std::vector<std::string> list_of_files;
-  for (int ii = 1; ii<list_of_files_tmp.size(); ii++) {
+  for (size_t ii = 1; ii<list_of_files_tmp.size(); ii++) {
     list_of_files.push_back(list_of_files_tmp[ii]);
   }
   list_of_files.push_back(list_of_files_tmp[0]);
